pyproject.toml
[tool.poetry]
name = "rag-system"
version = "0.1.0"
description = "RAG System"
authors = ["kioharu"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.11.*"
fastapi = { extras = ["standard"], version = "0.115.7" }
gunicorn = { extras = ["gevent"], version = "23.0.0" }
redis = { extras = ["hiredis"], version = "6.1.0" }
pydantic-ai-slim = { extras = ["google", "openai", "duckduckgo"], version = "1.60.0" }
logfire = {extras = ["fastapi", "redis", "sqlalchemy"], version = "4.24.0"}
qdrant-client = { extras = ["fastembed"], version = "1.17.0" }
sentry-sdk = "2.53.0"
alembic = "1.14.1"
pydantic = "2.11.7"
pydantic-settings = "2.10.1"
sqlalchemy = "2.0.37"
asyncpg = "0.28.0"
prefect = "3.6.17"
httpx = "0.28.1"
streamlit = "1.49.1"
pypdf = "5.6.0"
langchain-text-splitters = "0.3.11"
python-multipart = "0.0.20"
cryptography = "44.0.1"
python-docx = "1.2.0"
striprtf = "0.0.29"
odfpy = "1.4.1"
ebooklib = "0.19"
beautifulsoup4 = "4.13.4"
python-pptx = "1.0.2"
openpyxl = "3.1.5"
clickhouse-connect = "0.9.2"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pre-commit = "*"
ruff = "*"
ty = "*"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "*"
pytest-asyncio = "*"
pytest-cov = "*"
factory-boy = "*"
faker = "*"
testcontainers = { extras = ["postgresql"], version = "*" }

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "F",    # Pyflakes
    "E",    # pycodestyle
    "W",    # pycodestyle
    "C90",  # mccabe
    "I",    # isort
    "N",    # pep8-naming
    "S",    # flake8-bandit
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "EM",   # flake8-errmsg
    "ISC",  # flake8-implicit-str-concat
    "ICN",  # flake8-import-conventions
    "LOG",  # flake8-logging
    "G",    # flake8-logging-format
    "PIE",  # flake8-pie
    "T20",  # flake8-print
    "PT",   # flake8-pytest-style
    "RSE",  # flake8-raise
    "RET",  # flake8-return
    "SIM",  # flake8-simplify
    "TC",   # flake8-type-checking
    "PTH",  # flake8-use-pathlib
    "PGH",  # pygrep-hooks
    "PL",   # Pylint
    "TRY",  # tryceratops
    "PERF", # Perflint
    "FURB", # refurb
    "FAST", # FastAPI
]
exclude = ["db/alembic/versions/*.py", "db/alembic/env.py", "ai/prompts/*.py"]

[tool.ruff.lint.per-file-ignores]
"*.py" = [
    "C417",    # "Use map instead of list comprehension"
    "PLR0913", # "Too many arguments in function definition"
]
"tests/*.py" = [
    "S101",    # "Use of `assert` detected"
    "S107",    # "Possible hardcoded password assigned to function default"
    "PGH003",  # "Use specific rule codes when ignoring type issues"
]

[tool.ty.environment]
python-version = "3.11"

[tool.ty.src]
exclude = ["db/alembic/"]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::sqlalchemy.exc.MovedIn20Warning",
    "ignore::pydantic._internal._config.PydanticDeprecatedSince20"
]

[tool.coverage.run]
omit = ["tests/*", "db/alembic/versions/*"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
